requirements.txt
streamlit
torch
torchvision
opencv-python-headless
Pillow
scikit-image
numpy
